scheduler: pull tenant-querier assignments into separate file
